Ревю 2: Исправил передачу параметров формы, а также отделил App от корневого index
